refactor(view): 切り出した型が利用されるように修正
